src/requirements.txt
fastapi
uvicorn
scikit-learn
google-cloud-storage
google-cloud-aiplatform[prediction]
